refactor: enums 패키지 분리
