--- src/main/java/com/Ureka/AnalDoo/domain/entity/PayMethod.java
@@ -1,5 +1,0 @@
-package com.Ureka.AnalDoo.domain.entity;
-
-public enum PayMethod {
-    CARD
-}
