+++ src/main/java/com/Ureka/AnalDoo/domain/entity/Payment.java
@@ -1,6 +1,8 @@
 package com.Ureka.AnalDoo.domain.entity;
 
 import com.Ureka.AnalDoo.common.domain.TimeBaseEntity;
+import com.Ureka.AnalDoo.domain.entity.enums.PayMethod;
+import com.Ureka.AnalDoo.domain.entity.enums.PaymentStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
