--- src/main/java/com/Ureka/AnalDoo/domain/entity/SocialType.java
@@ -1,5 +1,0 @@
-package com.Ureka.AnalDoo.domain.entity;
-
-public enum SocialType {
-    KAKAO, GOOGLE
-}
