+++ src/main/java/com/Ureka/AnalDoo/domain/entity/enums/CompetitionStatus.java
@@ -1,15 +1,13 @@
-package com.Ureka.AnalDoo.domain.entity;
+package com.Ureka.AnalDoo.domain.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum Local {
-
-    SEOUL("서울"),GWANGJU("광주"),DAEGU("대구");
+public enum CompetitionStatus {
+    OPEN("모집중"),
+    CLOSED("모집마감");
 
     private final String name;
-
-
 }
