+++ src/main/java/com/Ureka/AnalDoo/domain/entity/enums/Local.java
@@ -1,13 +1,15 @@
-package com.Ureka.AnalDoo.domain.entity;
+package com.Ureka.AnalDoo.domain.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum CompetitionStatus {
-    OPEN("모집중"),
-    CLOSED("모집마감");
+public enum Local {
+
+    SEOUL("서울"),GWANGJU("광주"),DAEGU("대구");
 
     private final String name;
+
+
 }
