+++ src/main/java/com/Ureka/AnalDoo/domain/entity/enums/PayMethod.java
@@ -1,0 +1,5 @@
+package com.Ureka.AnalDoo.domain.entity.enums;
+
+public enum PayMethod {
+    CARD
+}
