+++ src/main/java/com/Ureka/AnalDoo/domain/entity/enums/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.Ureka.AnalDoo.domain.entity;
+package com.Ureka.AnalDoo.domain.entity.enums;
 
 public enum PaymentStatus {
     PAID,READY,CANCEL
