+++ src/main/java/com/Ureka/AnalDoo/domain/entity/enums/SocialType.java
@@ -1,0 +1,5 @@
+package com.Ureka.AnalDoo.domain.entity.enums;
+
+public enum SocialType {
+    KAKAO, GOOGLE
+}
