+++ src/main/java/com/Ureka/AnalDoo/domain/entity/enums/SportType.java
@@ -1,4 +1,4 @@
-package com.Ureka.AnalDoo.domain.entity;
+package com.Ureka.AnalDoo.domain.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
