+++ src/main/java/com/Ureka/AnalDoo/domain/payment/repository/PaymentRepository.java
@@ -1,10 +1,8 @@
 package com.Ureka.AnalDoo.domain.payment.repository;
 
-import com.Ureka.AnalDoo.domain.entity.Competition;
 import com.Ureka.AnalDoo.domain.entity.Payment;
-import com.Ureka.AnalDoo.domain.entity.PaymentStatus;
+import com.Ureka.AnalDoo.domain.entity.enums.PaymentStatus;
 import com.Ureka.AnalDoo.domain.entity.Reservation;
-import com.Ureka.AnalDoo.domain.entity.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
