+++ src/main/java/com/Ureka/AnalDoo/domain/payment/service/PortOnePaymentServiceImpl.java
@@ -3,9 +3,9 @@ package com.Ureka.AnalDoo.domain.payment.service;
 import com.Ureka.AnalDoo.common.exception.RestApiException;
 import com.Ureka.AnalDoo.common.exception.errorcode.PaymentErrorCode;
 import com.Ureka.AnalDoo.common.exception.errorcode.ReservationErrorCode;
-import com.Ureka.AnalDoo.domain.entity.PayMethod;
+import com.Ureka.AnalDoo.domain.entity.enums.PayMethod;
 import com.Ureka.AnalDoo.domain.entity.Payment;
-import com.Ureka.AnalDoo.domain.entity.PaymentStatus;
+import com.Ureka.AnalDoo.domain.entity.enums.PaymentStatus;
 import com.Ureka.AnalDoo.domain.entity.Reservation;
 import com.Ureka.AnalDoo.domain.entity.User;
 import com.Ureka.AnalDoo.domain.payment.dto.PaymentCancelResponse;
