+++ src/test/java/com/Ureka/AnalDoo/fixture/CompetitionFixture.java
@@ -3,9 +3,9 @@ package com.Ureka.AnalDoo.fixture;
 import com.Ureka.AnalDoo.domain.entity.Address;
 import com.Ureka.AnalDoo.domain.entity.Competition;
 import com.Ureka.AnalDoo.domain.entity.CompetitionPeriod;
-import com.Ureka.AnalDoo.domain.entity.CompetitionStatus;
-import com.Ureka.AnalDoo.domain.entity.Local;
-import com.Ureka.AnalDoo.domain.entity.SportType;
+import com.Ureka.AnalDoo.domain.entity.enums.CompetitionStatus;
+import com.Ureka.AnalDoo.domain.entity.enums.Local;
+import com.Ureka.AnalDoo.domain.entity.enums.SportType;
 import com.Ureka.AnalDoo.domain.entity.User;
 import java.time.LocalDateTime;
 import org.springframework.test.util.ReflectionTestUtils;
